Move AccessPolicy from access.py to project/enums.py.
