+++ physionet-django/project/modelcomponents/access.py
@@ -1,6 +1,5 @@
 import unicodedata
 from datetime import timedelta
-from enum import IntEnum
 
 from django.contrib.auth.hashers import check_password, make_password
 from django.contrib.contenttypes.fields import GenericForeignKey
@@ -12,23 +11,8 @@ from django.utils.crypto import get_random_string
 from project.fields import SafeHTMLField
 from project.validators import validate_version
 
+from project.enums import AccessPolicy
 from project.managers.access import DataAccessRequestQuerySet, DataAccessRequestManager
-
-
-class AccessPolicy(IntEnum):
-    OPEN = 0
-    RESTRICTED = 1
-    CREDENTIALED = 2
-    CONTRIBUTOR_REVIEW = 3
-
-    do_not_call_in_templates = True
-
-    @classmethod
-    def choices(cls, gte_value=0):
-        return tuple(
-            (option.value, option.name.replace("_", " ").title())
-            for option in cls if option.value >= gte_value
-        )
 
 
 class DUASignature(models.Model):
